+++ src/PhpPresentation/Writer/PowerPoint2007/PptCharts.php
@@ -925,6 +925,11 @@ class PptCharts extends AbstractDecoratorWriter
             // c:showLeaderLines
             $this->writeElementWithValAttribute($objWriter, 'c:showLeaderLines', $series->hasShowLeaderLines() ? '1' : '0');
 
+            // c:separator
+            if ($series->hasShowSeparator()) {
+                $this->writeElementWithValAttribute($objWriter, 'c:separator', $series->getSeparator());
+            }
+
             $objWriter->endElement();
 
             // c:spPr
@@ -1133,6 +1138,11 @@ class PptCharts extends AbstractDecoratorWriter
             // c:showLeaderLines
             $this->writeElementWithValAttribute($objWriter, 'c:showLeaderLines', $series->hasShowLeaderLines() ? '1' : '0');
 
+            // c:separator
+            if ($series->hasShowSeparator()) {
+                $this->writeElementWithValAttribute($objWriter, 'c:separator', $series->getSeparator());
+            }
+
             $objWriter->endElement();
 
             // c:spPr
@@ -1339,6 +1349,11 @@ class PptCharts extends AbstractDecoratorWriter
             // c:showLeaderLines
             $this->writeElementWithValAttribute($objWriter, 'c:showLeaderLines', $series->hasShowLeaderLines() ? '1' : '0');
 
+            // c:separator
+            if ($series->hasShowSeparator()) {
+                $this->writeElementWithValAttribute($objWriter, 'c:separator', $series->getSeparator());
+            }
+
             $objWriter->endElement();
 
             // Write X axis data
@@ -1504,6 +1519,11 @@ class PptCharts extends AbstractDecoratorWriter
             // c:showLeaderLines
             $this->writeElementWithValAttribute($objWriter, 'c:showLeaderLines', $series->hasShowLeaderLines() ? '1' : '0');
 
+            // c:separator
+            if ($series->hasShowSeparator()) {
+                $this->writeElementWithValAttribute($objWriter, 'c:separator', $series->getSeparator());
+            }
+
             $objWriter->endElement();
 
             // Write X axis data
@@ -1643,6 +1663,11 @@ class PptCharts extends AbstractDecoratorWriter
 
             // c:showLeaderLines
             $this->writeElementWithValAttribute($objWriter, 'c:showLeaderLines', $series->hasShowLeaderLines() ? '1' : '0');
+
+            // c:separator
+            if ($series->hasShowSeparator()) {
+                $this->writeElementWithValAttribute($objWriter, 'c:separator', $series->getSeparator());
+            }
 
             // > c:dLbls
             $objWriter->endElement();
@@ -1821,6 +1846,11 @@ class PptCharts extends AbstractDecoratorWriter
 
             // c:showLeaderLines
             $this->writeElementWithValAttribute($objWriter, 'c:showLeaderLines', $series->hasShowLeaderLines() ? '1' : '0');
+
+            // c:separator
+            if ($series->hasShowSeparator()) {
+                $this->writeElementWithValAttribute($objWriter, 'c:separator', $series->getSeparator());
+            }
 
             $objWriter->endElement();
 
@@ -2024,7 +2054,12 @@ class PptCharts extends AbstractDecoratorWriter
 
         // c:majorTickMark
         $objWriter->startElement('c:majorTickMark');
-        $objWriter->writeAttribute('val', 'none');
+        $objWriter->writeAttribute('val', $oAxis->getMajorTickMark());
+        $objWriter->endElement();
+
+        // c:minorTickMark
+        $objWriter->startElement('c:minorTickMark');
+        $objWriter->writeAttribute('val', $oAxis->getMinorTickMark());
         $objWriter->endElement();
 
         // c:tickLblPos
@@ -2066,9 +2101,7 @@ class PptCharts extends AbstractDecoratorWriter
 
         // Font - a:solidFill
         $objWriter->startElement('a:solidFill');
-
         $this->writeColor($objWriter, $oAxis->getFont()->getColor());
-
         $objWriter->endElement();
 
         // Font - a:latin
@@ -2147,6 +2180,20 @@ class PptCharts extends AbstractDecoratorWriter
                 $objWriter->writeAttribute('val', 'between');
             }
             $objWriter->endElement();
+
+            // c:majorUnit
+            if ($oAxis->getMajorUnit() != null) {
+                $objWriter->startElement('c:majorUnit');
+                $objWriter->writeAttribute('val', $oAxis->getMajorUnit());
+                $objWriter->endElement();
+            }
+
+            // c:minorUnit
+            if ($oAxis->getMinorUnit() != null) {
+                $objWriter->startElement('c:minorUnit');
+                $objWriter->writeAttribute('val', $oAxis->getMinorUnit());
+                $objWriter->endElement();
+            }
         }
 
         $objWriter->endElement();