+++ CHANGELOG.md
@@ -18,9 +18,13 @@
 - PowerPoint97 Reader : Support of Slide Note - @Progi1984 GH-226
 - PowerPoint2007 Reader : Support of Shape Table - @Progi1984 GH-240
 - PowerPoint2007 Reader : Support of Slide Note - @Progi1984 GH-226
+- PowerPoint2007 Reader : Support text direction in Alignment for Table - @Progi1984 GH-218
 - PowerPoint2007 Writer : Implement character spacing - @jvanoostrom GH-301
 - PowerPoint2007 Writer : Axis Bounds in Chart - @Progi1984 GH-269
-- PowerPoint2007 Writer : Implement Legend Key in Series for Chart  - @Progi1984 GH-319
+- PowerPoint2007 Writer : Implement Legend Key in Series for Chart - @Progi1984 GH-319
+- PowerPoint2007 Writer : Support text direction in Alignment for Table - @SeregPie GH-218
+- PowerPoint2007 Writer : Support tick mark & unit in Axis for Chart - @Faab GH-218
+- PowerPoint2007 Writer : Support separator in Series for Chart - @jphchaput GH-218
 - Misc : Added two methods for setting Border & Fill in Legend - @Progi1984 GH-265
 
 ## 0.7.0 - 2016-09-12